Fix for using a custom `source` directory
Paths are still passed around as relative. Code which needs to read files
(images, templates) and code which needs to manipulate paths
(Jekyll::ResponsiveImage::Utils#image_hash) has to be aware of the site source
path.

This got a bit messy in the end, but I couldn't find a better way without doing
a big refactor of the entire plugin.