using the ImageSize gem where it makes sense, to greatly speed up Jekyll builds for projects that do not use the 'auto_rotate' configuration
when the ResizeHandler figures out whether to resize or even generate an image from the original, using the ImageSize gem is way quicker than opening the image using MiniMagick. This makes for much shorter Jekyll build times.

Discarded alternative gems inlcude 'dimensions' and 'fastimage' because those take orientation into consideration when returning size. ImageMagick (or RMagick and MiniMagick) do not, as does the ImageSize gem. Hence the decision for the latter.